extracted_frames fixture moved to api_test.py
conftest.py will only contain shared pytest fixtures going forward. Fixture "extracted_frames" was not shared across multiple test files, therefore moved to the only test file that uses it.
