+++ tests/api_test.py
@@ -1,9 +1,25 @@
+import shutil
 import typing as t
+from pathlib import Path
 
 import pytest
 
 from videoxt.utils import parse_kwargs
 from videoxt.video import VideoProperties
+
+
+@pytest.fixture(scope="class")
+def extracted_frames(tmp_video_filepath: Path) -> t.Dict[str, t.Any]:
+    """Extract frames from a video and yield the directory where they were saved.
+
+    Yields:
+        Path: The directory where the frames were saved.
+    """
+    from videoxt.api import extract_frames
+
+    results = extract_frames(tmp_video_filepath)
+    yield results
+    shutil.rmtree(results["request"]["destdir"])
 
 
 @pytest.mark.usefixtures("extracted_frames", "video_properties")
